Cleanup composer.json for 0.0.x-dev
Removed 0.0.1-beta1 version as it was not actually doing anything. Added
a branch alias that will enable people to require:

    "ircmaxell/security-lib": "0.0.*@dev"

An unintended side effect of composer.json changes is that it results in
composer.lock needing to change to account for a different hash. In addition,
Composer sometimes actually changes other aspects of the composer.lock
structure.