Cleanup composer.json
The version tag is not required since the branch is named like 1.0
it will automatically receive the 1.0.x-dev treatment by Composer.

An unintended side effect of composer.json changes is that it results in
composer.lock needing to change to account for a different hash. In addition,
Composer sometimes actually changes other aspects of the composer.lock
structure.